app/requirements.txt
lightrag-hku
networkx
pyvis
openai
langchain